Fix webpack resolution of dependencies
When webpack imports dependencies it will look in the local node_modules
folder. The library troposphere-ui imports react, so webpack looks in it that
libraries node_modules for that version of react. React is also imported from
the guide. So with webpack's simple module resolution multiple copies of react
get bundled. The webpack resolve root config will adjust the search path such
that the root paths are searched first, this way we can ensure that the
guide's version of react will always be preferred.