Refactor StyleGuide to use SideBar
Refactored the StyleGuide to use a SideBar Component and a flexbox
Layout. Removed the left margin on main that accounted for the sidebar
and added a wrapper in SideBar that maintains the SideBar's width. Now
the main content doesn;t need be concerned if the sidebar is open or
closed.