Ensure internal capitalization present for CyVerse
The brand for organization incorporates an InterCap [0], also called
BiCapitalization [1]. We could say that within source code, a usage
of `<CyverseLogo/>` would be totally fine because a competing
conversation, programming language style/convention would be more
applicable. A situational example would be when referring to the theme,
`cyverseTheme` is correct because of the "camelCase" idiom of
JavaScript. When we "lint" for conventions, not using camelCase would be
a violation.

The modified text in this commit are all examples of present text
appearing in "display examples" or documentation. In situation where
source convention should be favored, <code> was used to indicate the
usage as code.

[0] http://catb.org/~esr/jargon/html/I/InterCaps.html
[1] http://catb.org/~esr/jargon/html/B/BiCapitalization.html