Add React Tap Event plugin
The Style Guide now leverages "react-tap-event-plugin" [0]. Add this
dependency to the package.json. This plugin gives "tap" events without
the "delay" seen in iOS devices [1]. As noted in the project README [1],
Facebook is not planning on fixing this within React as it expects the
browsers to _soon_ correct the problem.

[0] https://github.com/zilverline/react-tap-event-plugin
[1] https://github.com/zilverline/react-tap-event-plugin#introduction