Remove command to adjust directory location
When the deploy script was added, it was assumed that it would be run
from within `./guide/script`. This would mean that the script should
alter the working directory (moving up two directories). However, the
version of the Travis-CI build that is in place is running the script
from the root repository with the path _to_ the script:

```
$ ./guide/scripts/deploy-to-gh-pages.sh
```

This eliminates any **need** to call change-directory within the script.