Use example file for example code documentation
Before we were manually copy pasting our example source code into a string to render the accompanying code to our example. This caused the example code to often be out of date.

Now the code example is generated from the example source directly.

As a bonus syntax highlighting has been added.

CSS rules were added to to code class selectors that help with formatting our code elements since the code elements are generated by our markdown component. We will favor Components for code styling over CSS in the lib for consuming applications but for the guide in this case the CSS selectors are expectable. Once these components are added we might consider switching our instances here to use them for the sake of consistency.