Fix issues with Travis build
When Travis publishes our GH page the npm package is referenced
instead of the local lib source. This causes the GH page to fail because
the npm package doesn't contain the changes the guide depends on. Since
the guide use npm link to reference the local lib source rather than the
published npm package we don't see this problem locally.
This commit adds a webpack alias to the local lib source.

Also, `makeTheme.js` uses `material-ui` to build our theme.json since we
include `material-ui` peer-dependency it isn't included in `node_moduls` and fails.

Adds `material-ui` as a dev-dependency so that it still
doesn't bundle in the npm package but can use it for building.

Adds install and build scripts for lib to `.travis.yml`