Allow trailing commas, where valid in ES5
Using "trailingComma": "es5" means that it will only be allowed where it
is valid within the ES5 specification [0].

This would mean that trailing commas would *not* be allowed in the ES6
destructuring syntax (though, MDN document considers that valid [1]).

The reason I am suggesting the `"es5"` option is for consideration as
the source is bundled in compliant ES5 formats.

Arguments for "all" could be made with a pull request built on-top of
this work and considered in the spirit of C4, code-as-conversation.

[0] https://prettier.io/docs/en/options.html#trailing-commas
[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Trailing_commas